requirements.txt
pillow
opencv-python
pytest